# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HireForward — AI-Native Recruitment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /processes/{id}/manager-link:
    post:
      operationId: generateManagerLink
      tags: [manager]
      summary: Generate a read-only manager view link for a process
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Manager link generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagerLink"
        "401":
          description: Unauthorized
        "404":
          description: Not found

  # ── Manager View (public, read-only) ───────────────────────────────────────
  /manager/{token}:
    get:
      operationId: getManagerView
      tags: [manager]
      summary: Get manager view — process overview with candidate scores (public)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Manager view
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagerView"
        "404":
          description: Not found

  # ── Configurator Agent (chat to set up a process) ─────────────────────────
...
  /interview/{token}:
    get:
      operationId: getInterviewByToken
      tags: [interviews]
      summary: Get interview info by invite token (public)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Interview info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewInfo"
        "404":
          description: Invalid or expired token
  /interview/{token}/start:
    post:
      operationId: startInterview
      tags: [interviews]
      summary: Start the interview session
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartInterviewBody"
      responses:
        "200":
          description: Session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewSession"
  /interview/{token}/message:
    post:
      operationId: sendInterviewMessage
      tags: [interviews]
      summary: Send a message to the interviewer agent (SSE stream)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendInterviewMessageBody"
      responses:
        "200":
          description: SSE stream of interviewer agent response
          content:
            text/event-stream: {}
  /interview/{token}/end:
    post:
      operationId: endInterview
      tags: [interviews]
      summary: End the interview session and trigger evaluation
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Interview ended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewEndResult"

  # ── Evaluations ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    InterviewInfo:
      type: object
      properties:
        candidateId:
          type: string
        candidateName:
          type: string
        processTitle:
          type: string
        companyName:
          type: string
        companyLogoUrl:
          type: string
          nullable: true
        interviewType:
          type: string
        toolsAllowed:
          type: boolean
        responseTimeLimitSeconds:
          type: integer
          nullable: true
        estimatedMinutes:
          type: integer
        status:
          type: string
          enum: [pending, started, completed, expired]
      required: [candidateId, candidateName, processTitle, companyName, interviewType, toolsAllowed, estimatedMinutes, status]
    StartInterviewBody:
      type: object
      properties:
        candidateName:
          type: string
      required: [candidateName]
    InterviewSession:
      type: object
      properties:
        sessionId:
          type: string
        startedAt:
          type: string
          format: date-time
      required: [sessionId, startedAt]
    SendInterviewMessageBody:
      type: object
      properties:
        content:
          type: string
        sessionId:
          type: string
      required: [content, sessionId]
    InterviewEndResult:
      type: object
      properties:
        sessionId:
          type: string
        endedAt:
          type: string
          format: date-time
        evaluationStatus:
          type: string
          enum: [processing, completed]
      required: [sessionId, endedAt, evaluationStatus]
...
    CriterionScore:
      type: object
      properties:
        name:
          type: string
        score:
          type: integer
        max:
          type: integer
        weight:
          type: integer
        justification:
          type: string
      required: [name, score, max, weight, justification]
...
    ManagerLink:
      type: object
      properties:
        managerViewToken:
          type: string
        link:
          type: string
      required: [managerViewToken, link]
    ManagerCandidate:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [invited, started, completed, expired]
        overallScore:
          type: integer
          nullable: true
        recommendation:
          type: string
          nullable: true
        summary:
          type: string
          nullable: true
        highlights:
          type: array
          nullable: true
          items:
            type: string
        redFlags:
          type: array
          nullable: true
          items:
            type: string
        criteriaScores:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/CriterionScore"
        createdAt:
          type: string
          format: date-time
      required: [id, name, status, createdAt]
    ManagerView:
      type: object
      properties:
        processId:
          type: string
        processTitle:
          type: string
        companyName:
          type: string
        companyLogoUrl:
          type: string
          nullable: true
        area:
          type: string
        seniority:
          type: string
        candidates:
          type: array
          items:
            $ref: "#/components/schemas/ManagerCandidate"
      required: [processId, processTitle, companyName, area, seniority, candidates]
...