# lib/api-zod/src/generated/types/interviewInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HireForward — AI-Native Recruitment Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { InterviewInfoStatus } from "./interviewInfoStatus";

export interface InterviewInfo {
  candidateId: string;
  candidateName: string;
  processTitle: string;
  companyName: string;
  companyLogoUrl?: string | null;
  interviewType: string;
  toolsAllowed: boolean;
  responseTimeLimitSeconds?: number | null;
  estimatedMinutes: number;
  status: InterviewInfoStatus;
}