# lib/api-zod/src/generated/types/managerView.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HireForward — AI-Native Recruitment Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ManagerCandidate } from "./managerCandidate";

export interface ManagerView {
  processId: string;
  processTitle: string;
  companyName: string;
  companyLogoUrl?: string | null;
  area: string;
  seniority: string;
  candidates: ManagerCandidate[];
}