# lib/api-zod/src/generated/types/sendInterviewMessageBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HireForward — AI-Native Recruitment Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface SendInterviewMessageBody {
  content: string;
  sessionId: string;
}